new ta queue for opening and closing office hours

+++ bot/command.py
@@ -403,6 +403,9 @@ class QueueStatus(Command):
 class AcceptStudent(Command):
     async def handle(self):
         qa: QueueAuthority = QueueAuthority(self.guild)
+        if not qa.is_ta_on_duty(self.message.author.id):
+            await self.message.channel.send("You must be on duty to accept a request!")
+            return
         # get oldest queue item (and also remove it)
         session: OHSession = await qa.dequeue(self.message.author)
 
@@ -494,7 +497,7 @@ class RejectStudent(Command):
                 malformed = False
 
                 # Check to see if a there is a valid member id and message
-                guild: Guild = message.guild;
+                guild: Guild = message.guild
                 try:
                     # checks for at least 3 words
                     if len(message.content.split(" ")) < 3:
@@ -625,13 +628,18 @@ async def is_oh_command(client, message, type):
 class StartOfficeHours(Command):
     async def handle(self):
         qa: QueueAuthority = QueueAuthority(self.guild)
-        qa.open_office_hours()
         ca: ChannelAuthority = ChannelAuthority(self.guild)
-        await ca.waiting_channel.send(
-            "Office hours are live.  Get in line with !request")
-        logger.info("Office hours opened by {}".format(
-            name(self.message.author)
-        ))
+        # Add TA to list of available TAs. Return True if fresh opening, True if newly added TA
+        fresh_open, is_new_ta = qa.open_office_hours(self.message.author.id)
+        if fresh_open:
+            await ca.queue_channel.send("Office hours are live!")
+            await ca.waiting_channel.send(
+                "Office hours are live.  Get in line with !request")
+            logger.info("Office hours opened by {}".format(
+                name(self.message.author)
+            ))
+        if is_new_ta:
+            await ca.queue_channel.send(name(self.message.author) + " has checked into office hours!")
 
     @staticmethod
     async def is_invoked_by_message(message: Message, client: Client):
@@ -642,13 +650,20 @@ class StartOfficeHours(Command):
 class EndOfficeHours(Command):
     async def handle(self):
         qa: QueueAuthority = QueueAuthority(self.guild)
-        qa.remove_all()
         ca: ChannelAuthority = ChannelAuthority(self.guild)
-        await ca.waiting_channel.send(
-            "Ok, y'all.  Office hours have ended for now.  An announcement will appear here when they have reopened.")
-        logger.info("Office hours closed by {}".format(
-            name(self.message.author)
-        ))
+        # Remove TA from list of available TAs.
+        is_open, was_removed, ta_count = qa.close_office_hours(self.message.author.id)
+        if ta_count > 0 and is_open and was_removed:
+            await ca.queue_channel.send(name(self.message.author) + " has checked out of office hours.")
+        # Last TA was removed, or TA queue was empty when called
+        elif ta_count <= 0 and is_open:
+            await qa.remove_all()
+            await ca.queue_channel.send(name(self.message.author) + " has closed office hours!")
+            await ca.waiting_channel.send(
+                "Ok, y'all.  Office hours have ended for now.  An announcement will appear here when they have reopened.")
+            logger.info("Office hours closed by {}".format(
+                name(self.message.author)
+            ))
 
     @staticmethod
     async def is_invoked_by_message(message: Message, client: Client):
